pyproject.toml
[project]
name = "handmark"
version = "0.4.0"
description = "Transform handwritten images into structured documents (Markdown, JSON, YAML, XML)"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Gabriel Borges", email = "dev.gabrielsborges@proton.me"}
]
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "pytest",
    "pytest-mock",
    "typer>=0.15.4",
    "rich>=14.0.0",
    "pyyaml>=6.0",
    "lxml>=4.9.0",
]

[project.urls]
"Homepage" = "https://github.com/devgabrielsborges/handmark"
"Bug Tracker" = "https://github.com/devgabrielsborges/handmark/issues"

[project.scripts]
handmark = "main:main"

[dependency-groups]
dev = [
    "flake8>=7.3.0",
]
